Handle fees and add GetAccount api
This adds many changes surrounding configuring the functions to handle
fees from both the venues and the Neoworks side (retail). The way this
works currently is that these fields are optional to allow for
OrderManager to begin sending the values. When no value is found for
fees they will be set to 0 for the function execution, preventing an
entry from being written. Otherwise the fees will get their own entries
and then will be contributing to final balance calculations.

For the GetAccount, the api is configured to pull the account details
and the latest balance and return both back to the user.

This has an additional change where the Demo and Demo1 accounts are
initialized. The balances will be set in accordance with how the demo
app will run.