Change idempotency in partial release to check for entry
This changes the check for the idempotency to look for a corresponding
fill entry with the request id passed as argument. Right now this is the
fill id (which should be static for each fill) and will let us identify
if the release already happened.

Previously holds were used, but given that a partial release could be
the final fill, no additional hold would be placed. Checking for the
sender entry corresponding to the fill will allow us to ensure that we
are not going to miss a fill and double ledger.