Add ErrorRepository and chain repository interfaces
This adds the ErrorRepository type, which wraps all errors so the passed
back message is user readable and not filled with implementation details
like a Scan failure. The intention of this is to make error handling
more streamlined for the OrderManager and frontend. This was tested
using a new error example.

This also includes a cleanup of the main function, moving it to
cmd/server, unifying with OrderManager on how the repo is structured.